% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_word.R
\name{hello_word}
\alias{hello_word}
\title{hello_word}
\usage{
hello_word(what)
}
\arguments{
\item{df}{text to be returned}
}
\description{
return text
}
\examples{
hello_world()
}
\keyword{hello}
\keyword{world}
