% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dendrometer_data.R
\name{load_dendrometer_data}
\alias{load_dendrometer_data}
\title{load_dendrometer_data}
\usage{
load_dendrometer_data(data_folder, output_format = "data.frame")
}
\arguments{
\item{data_folder}{input data folder}

\item{definies}{output format (list or data.frame)}
}
\description{
return data.frame
}
\examples{
load_dendrometer_data()
}
\keyword{data}
\keyword{load}
